feat: add new js file for keyboardLayouts
