+++ index.html
@@ -11,7 +11,7 @@
 </head>
 
 <body>
-	<script src="./js/index.js"></script>
+	<script type="module" src="./js/index.js"></script>
 </body>
 
 </html>
