+++ js/keys.js
@@ -1,0 +1,108 @@
+const keyboardLayouts = {
+  en: [
+    [
+      '`',
+      '1',
+      '2',
+      '3',
+      '4',
+      '5',
+      '6',
+      '7',
+      '8',
+      '9',
+      '0',
+      '-',
+      '=',
+      'Backspace',
+    ],
+    [
+      'Tab',
+      'q',
+      'w',
+      'e',
+      'r',
+      't',
+      'y',
+      'u',
+      'i',
+      'o',
+      'p',
+      '[',
+      ']',
+      '\\',
+      'Del',
+    ],
+    [
+      'Caps lock',
+      'a',
+      's',
+      'd',
+      'f',
+      'g',
+      'h',
+      'j',
+      'k',
+      'l',
+      ';',
+      "'",
+      'Enter',
+    ],
+    ['Shift', 'z', 'x', 'c', 'v', 'b', 'n', 'm', ',', '.', '/', '▲', 'Shift'],
+    ['Ctrl', 'Win', 'Alt', 'space', 'Alt', '◄', '▼', '►', 'Ctrl'],
+  ],
+  ru: [
+    [
+      'ё',
+      '1',
+      '2',
+      '3',
+      '4',
+      '5',
+      '6',
+      '7',
+      '8',
+      '9',
+      '0',
+      '-',
+      '=',
+      'Backspace',
+    ],
+    [
+      'Tab',
+      'й',
+      'ц',
+      'у',
+      'к',
+      'е',
+      'н',
+      'г',
+      'ш',
+      'щ',
+      'з',
+      'х',
+      'ъ',
+      '\\',
+      'Del',
+    ],
+    [
+      'Caps lock',
+      'ф',
+      'ы',
+      'в',
+      'а',
+      'п',
+      'р',
+      'о',
+      'л',
+      'д',
+      'ж',
+      'э',
+      'Enter',
+    ],
+    ['Shift', 'я', 'ч', 'с', 'м', 'и', 'т', 'ь', 'б', 'ю', '.', '▲', 'Shift'],
+    ['Ctrl', 'Win', 'Alt', 'space', 'Alt', '◄', '▼', '►', 'Ctrl'],
+  ],
+};
+
+export default keyboardLayouts;
