+++ package.json
@@ -1,5 +1,4 @@
 {
-	"type": "module",
 	"devDependencies": {
 		"eslint": "^8.39.0",
 		"eslint-config-airbnb-base": "^15.0.0",
